FuzzyCRegression.aic — Method

aic()

Calculates Aikike Inforation Criteria for fitted FCR model, used for selecting optimal number of groups

FuzzyCRegression.bic — Method

bic()

Calculates Bayesian Inforation Criteria for fitted FCR model, used for selecting optimal number of groups

FuzzyCRegression.coef — Method

coef()

Extract coefficients from fitted model

FuzzyCRegression.coefnames — Method

coefnames()

Returns names of coefficients from fitted model

FuzzyCRegression.confint — Method

confint()

Returns lower and upper confidence interval for fitted model, for specified significance level (default = 0.95)

FuzzyCRegression.distribution — Method

distribution()

Calculates distribution of weighted coefficients from fitted model

Arguments

results::FCRModel Model type from fcr output

index::Integer Column index of variable in X matrix to calculate coefficient distibution for (defaults to 1)

FuzzyCRegression.fit — Method

fit()

Fits the FCR model

Arguments

- `df`: name of dataframe (if missing, data must be passed as arrays)
- `y`: column name or array holding values of the dependent variable (required)
- `X`: a list of column names or a matrix holding values of the independent variable(s) with heterogeneous coefficients (required)
- `Z`: a list of column names or a matrix holding values of the independent variable(s) with homogeneous coefficients
- `G`: number of groups (required)
- `m`: regularization parameter (greater than 1), where group assignment becomes binary as m approaches 1 (default = 1.5)
- `unit`: column name or array with unit identifier (if panel structure)
- `time`: column name or array with time indicators (if panel structure)
- `startvals`: number of starting values for the minimization routine (default = 100)
- `cores`: number of parallel workers (default = 1)

FuzzyCRegression.predict — Method

predict()

Obtain predicted values of the dependent variable from the fitted model, using modal group membership

FuzzyCRegression.residuals — Method

residuals()

Get the vector of residuals from the fitted model

FuzzyCRegression.stderror — Method

stderror()

Standard errors from fitted model

FuzzyCRegression.summarize — Method

summarize()

Summarizes results from fitted model

FuzzyCRegression.vcov — Method

vcov()

Variance covariance matrix from fitted model

FuzzyCRegression.weights — Method

weights()

Calculate group weights from fitted model, using modal group membership