FuzzyCRegression.aic — Method

aic()

Calculates Aikike Inforation Criteria for fitted FCR model, used for selecting optimal number of groups

FuzzyCRegression.bic — Method

bic()

Calculates Bayesian Inforation Criteria for fitted FCR model, used for selecting optimal number of groups

FuzzyCRegression.coef — Method

coef()

Extract coefficients from fitted model

FuzzyCRegression.coefnames — Method

coefnames()

Returns names of coefficients from fitted model

FuzzyCRegression.distribution — Method

distribution()

Calculates distribution of weighted coefficients from fitted model

Arguments

results::Model Model type from fcr output

index::Integer Column index of variable in X matrix to calculate coefficient distibution for (defaults to 1)

FuzzyCRegression.fit — Method

fit()

Implements fuzzy clustering regression estimator from Lewis, Melcangi, Pilossoph, and Toner-Rodgers (2022)

Arguments

y::Vector dependent var

x::Matrix variables with heterogeneous coefficients (defaults to vector of 1's for FEs)

Z:matrix matrix of controls

unit:Vector vector of unit IDs

t::Vector time vector (optional)

G::Integer number of groups (default = 2)

m::Real fuzzy tuning parameter

startvals::Integer number of starting values (default = 1,000)

cores::Integer number of cores (default = 1)

Returns:

struct with the following methods:

coefficients::Vector: vector of coefficients

weights::Matrix: group weights

SE::Vector: standard errors (optional)

objective::Number value of objective function at minimum

FuzzyCRegression.predict — Method

predict()

Obtain predicted values of the dependent variable from the fitted model, using modal group membership

FuzzyCRegression.residuals — Method

residuals()

Get the vector of residuals from the fitted model

FuzzyCRegression.stderror — Method

stderror()

Standard errors from fitted model

FuzzyCRegression.summarize — Method

summarize()

Summarizes results from fitted model

FuzzyCRegression.vcov — Method

vcov()

Variance covariance matrix from fitted model

FuzzyCRegression.weights — Method

weights()

Calculate group weights from fitted model, using modal group membership