FuzzyCRegression.distribution — Method

distribution()

Calculates distribution of weighted coefficients for fcr model output

Arguments

results::Model Model type from fcr output

index::Integer Column index of variable in X matrix to calculate coefficient distibution for (defaults to 1)

plot:Bool Plot histogram of coefficients

FuzzyCRegression.fit — Method

fit()

Implements fuzzy clustering regression estimator from Lewis, Melcangi, Pilossoph, and Toner-Rodgers (2022)

Arguments

y::Vector dependent var

x::Matrix variables with heterogeneous coefficients (defaults to vector of 1's for FEs)

Z:matrix matrix of controls

unit:Vector vector of unit IDs

t::Vector time vector (optional)

G::Integer number of groups (default = 2)

m::Real fuzzy tuning parameter

startvals::Integer number of starting values (default = 1,000)

cores::Integer number of cores (default = 1)

Returns:

struct with the following methods:

coefficients::Vector: vector of coefficients

weights::Matrix: group weights

SE::Vector: standard errors (optional)

objective::Number value of objective function at minimum

FuzzyCRegression.information — Method

information()

Calculates inforation criteria for fcr model output, used for selecting optimal number of groups

Arguments

results::Model Model type from fcr output

criterion::String Information criterion (AIC, BIC, or HQC)